version1

+++ paseos/views.py
@@ -1,5 +1,5 @@
 from django.http import HttpResponse
-from django.shortcuts import redirect, render
+from django.shortcuts import redirect, render, get_object_or_404
 from paseos.backends import AdminAuthentication
 from django.template import loader
 from django.core.exceptions import ObjectDoesNotExist
@@ -218,7 +218,7 @@ def verPaseosAdmin(request):
 def paseoAdmin(request, id):
     try:
         paseo = Paseo.objects.get(pk=id)
-        chiva = Chiva.objects.get(pk=paseo.chiva.placa)
+        chiva = Chiva.objects.get(pk=paseo.chiva.id)
         esquema = EsquemaCobro.objects.get(pk=paseo.esquemaCobro.id)
         listaReservas = paseo.reserva_set.all()
 
@@ -235,8 +235,26 @@ def paseoAdmin(request, id):
     return HttpResponse(template.render(context, request))
 
 def chivas(request):
+    listaChivas = Chiva.objects.all()
+
     if request.method == 'POST':
         form = ChivaForm(request.POST)
+        chivaPlaca = request.POST.get("placa")
+        chivaCapacidad = request.POST.get("capacidad")
+        chivaTipo = request.POST.get("tipo")
+        chivaEstado = request.POST.get("estado")
+        
+        print(type(chivaCapacidad))
+
+
+       
+        if (chivaTipo == 'Normal') or (chivaTipo =='Rumbera'):
+            pass
+        else:
+            messages.error(request, 'Tipo de chiva no válido.')
+            return render(request, 'chivas.html', {'listaChivas': listaChivas, 'chivaPlaca' : chivaPlaca, 'chivaCapacidad' : chivaCapacidad })
+        
+
         if form.is_valid():
             form.save()
             messages.success(request, 'Chiva guardada con exito')
@@ -244,22 +262,68 @@ def chivas(request):
     else:
         form = ChivaForm()
 
-    listaChivas = Chiva.objects.all()
+
     template = loader.get_template('chivas.html')
     context = {
         'listaChivas': listaChivas,
         'form': form,
+        'actualización': None
     }
     return HttpResponse(template.render(context, request))
 
-def eliminarChiva(request, placa):
+def actualizarFormChiva(request, id):
+    chiva = get_object_or_404(Chiva, pk=id)
+
+    if request.method == 'POST':
+        chivaPlaca = request.POST.get("placa")
+        chivaCapacidad = request.POST.get("capacidad")
+        print( request.POST.get("tipo"))
+        chivaTipo = request.POST.get("tipo")
+        chivaEstado = request.POST.get("estado")
+        
+        chiva.placa = chivaPlaca
+        chiva.capacidad = chivaCapacidad
+        chiva.tipo = chivaTipo
+        chiva.estado = chivaEstado
+        chiva.save()
+
+        messages.success(request, 'Chiva actualizada con éxito')
+        return redirect('chivas')
+    else:
+        form = ChivaForm(instance=chiva)
+
+    paseoChiva = None
     try:
-        chiva = Chiva.objects.get(pk=placa)
+        paseoChiva = Paseo.objects.get(chiva_id=id)
+        print(paseoChiva.id)
+    except ObjectDoesNotExist:
+        pass
+        
+        # messages.error(request,"Por favor seleccione una chiva válida.")
+    
+
+    listaChivas = Chiva.objects.all()
+
+    context = {
+        'listaChivas': listaChivas,
+        'form': form,
+        'actualizacion': chiva,
+        'paseoChiva': paseoChiva
+    }
+    return render(request, 'chivas.html', context)
+
+def eliminarChiva(request, id):
+    try:
+        chiva = Chiva.objects.get(pk=id)
     except Chiva.DoesNotExist:
         return messages.error('Chiva no encontrada')
     
-    chiva.delete()
-    messages.warning(request, 'Chiva eliminada con exito')
+    if Paseo.objects.filter(chiva_id=id).exists():
+        messages.error(request, 'Error al eliminar la chiva, esta se encuentra asociada a un paseo')
+    else:
+        chiva.delete()
+        messages.warning(request, 'Chiva eliminada con éxito')
+  
     return redirect('../')
 
 def hola(request):
